leak var ama .alışıyo

+++ utils/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: fkuyumcu <fkuyumcu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 11:42:52 by fkuyumcu          #+#    #+#             */
-/*   Updated: 2024/12/29 12:45:58 by fkuyumcu         ###   ########.fr       */
+/*   Updated: 2024/12/30 13:29:10 by fkuyumcu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int	i;
 	int	sign;
